fix for tests with valgrind

+++ src/s21_sprintf_funcs.h
@@ -68,11 +68,11 @@ long int get_power(long double num);
 void print_whole_float(long double num, char* dst);
 void print_fractional_float(long double fractional, int precision, int plus_sgn,
                             char* dst);
-int check_special_float_nums(long double num, int width, int right_padding,
+int check_special_float_nums(double num, int width, int right_padding,
                              char pading_symbol, int plus_sgn, int space_symbol,
                              char* dst);
 void print_inf_nan_sgn(int sgn, int plus_sgn, int space_symbol, char* dst);
-long double round_double(long double num, int* power, int precision);
+long double round_double(long double num, long int* power, int precision);
 void print_double_shortest(long double num, int precision, int width,
                            int right_padding, char pading_symbol, int plus_sgn,
                            int space_symbol, int capital, int point_forced,
